Check for empty strings on saving the position field settings
When the position field is used in a Matrix block setting, the `options` attribute receives `"1"` or `""` as the values for the options (left, right, center, etc). Until PHP 7.*, loosely comparing `""` to `0` returned `true`. As of PHP 8, the same comparison returns `false` (see https://www.php.net/manual/en/migration80.incompatible.php). The proposed change adds empty strings as a valid value for any of the `options`.